패키지 구조 전체 변경 - 패키지 구조가 아래와 같은지 검토 한번 해주시길 바랍니다.
com.review.StoreReview
					> StoreReviewApplication
com.review.StoreReview.common
com.review.StoreReview.common.enum
								> KeyNameDescEnum.java
								> ApiStatusCode.java
								> Gender.java
								> UseYN.java
com.review.StoreReview.config
							> CmsJpaConfig.java
							> CustJpaConfig.java
							> LogJpaConfig.java
							> MngJpaConfig.java
com.review.StoreReview.aop
						> RequestAroundLogAop.java
com.review.StoreReview.filter (예정)
-------------------------------------------------------------------------------------------------
com.review.StoreReview.controller
							> BaseController.java
com.review.StoreReview.controller.impl.cms
									> ~~~~Controller.java
com.review.StoreReview.controller.impl.cust
									> ~~~~Controller.java
com.review.StoreReview.controller.impl.log
									> ~~~~Controller.java
com.review.StoreReview.controller.impl.mng
									> ~~~~Controller.java

com.review.StoreReview.service
							> BaseService.java
com.review.StoreReview.service.impl.cms
									> ~~~~~Service.java
com.review.StoreReview.service.impl.cust
									> ~~~~~Service.java
com.review.StoreReview.service.impl.log
									> ~~~~~Service.java
com.review.StoreReview.service.impl.mng
									> ~~~~~Service.java

com.review.StoreReview.repository
							> BaseRepository.java
com.review.StoreReview.repository.impl.cms
									> ~~~~~Repository.java
com.review.StoreReview.repository.impl.cust
									> ~~~~~Repository.java
com.review.StoreReview.repository.impl.log
									> ~~~~~Repository.java
com.review.StoreReview.repository.impl.mng
									> ~~~~~Repository.java
-------------------------------------------------------------------------------------------------
com.review.StoreReview.dao(DataBase Access Object)
						> BaseTimeEntity.java
com.review.StoreReview.dao.cms
							> Commnet.java
							> Review.java
com.review.StoreReview.dao.cust
							> User.java
com.review.StoreReview.dao.log
							> Log.java
com.review.StoreReview.dao.mng
							> MngCode.java
							> MngCodeInfo.java
-------------------------------------------------------------------------------------------------
com.review.StoreReview.dto( Data Transfer Object )
com.review.StoreReview.dto.impl.request
com.review.StoreReview.dto.impl.response
