--- src/main/java/com/review/StoreReview/common/BaseTimeEntity.java
@@ -1,20 +1,0 @@
-package com.review.StoreReview.common;
-
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-import javax.persistence.Column;
-import javax.persistence.EntityListeners;
-import javax.persistence.MappedSuperclass;
-import java.time.LocalDateTime;
-
-@EntityListeners(AuditingEntityListener.class)      // Auditing 기능을 포함
-@MappedSuperclass   //이 class를 상속받으면 필드를 column으로 인식
-public abstract class BaseTimeEntity {
-    @CreatedDate
-    @Column(updatable = false)
-    private LocalDateTime createdAt;
-
-    @LastModifiedDate
-    private LocalDateTime modifiedAt;
-}
