--- src/main/java/com/review/StoreReview/common/impl/KeyValueEnumImpl.java
@@ -1,7 +1,0 @@
-package com.review.StoreReview.common.impl;
-
-public interface KeyValueEnumImpl {
-    Integer getKey();
-    String getName();
-    String getDesc();
-}
