--- src/main/java/com/review/StoreReview/domain/enums/Gender.java
@@ -1,4 +1,0 @@
-package com.review.StoreReview.domain.enums;
-
-public enum Gender {
-    W, M}
