--- src/main/java/com/review/StoreReview/domain/enums/UseYN.java
@@ -1,5 +1,0 @@
-package com.review.StoreReview.domain.enums;
-
-public enum UseYN {
-    Y, N
-}
