--- src/main/java/com/review/StoreReview/repository/UserRepository.java
@@ -1,9 +1,0 @@
-package com.review.StoreReview.repository;
-
-import com.review.StoreReview.domain.CUST.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, String> {   // SUID를 id로 지정.
-}
