--- src/main/java/com/review/StoreReview/service/UserServiceImpl.java
@@ -1,22 +1,0 @@
-package com.review.StoreReview.service;
-
-import com.review.StoreReview.repository.UserRepository;
-import com.review.StoreReview.web.rest.controller.dto.UserSaveRequestDto;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserServiceImpl implements UserService {
-    private final UserRepository userRepository;
-
-    @Autowired
-    public UserServiceImpl(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    @Override
-    public String join(UserSaveRequestDto requestDto) {
-        // 회원가입 로직
-        return userRepository.save(requestDto.toEntity()).getSuid();
-    }
-}
