--- src/main/java/com/review/StoreReview/web/rest/controller/dto/ResponseDto.java
@@ -1,12 +1,0 @@
-package com.review.StoreReview.web.rest.controller.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.NoArgsConstructor;
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class ResponseDto<T> {
-    int status;
-    T data;
-}
