+++ src/main/java/com/review/storereview/common/enumerate/Gender.java
@@ -1,0 +1,4 @@
+package com.review.storereview.common.enumerate;
+
+public enum Gender {
+    W, M}
