+++ src/main/java/com/review/storereview/common/enumerate/KeyValueEnumImpl.java
@@ -1,0 +1,7 @@
+package com.review.storereview.common.enumerate;
+
+public interface KeyValueEnumImpl {
+    Integer getKey();
+    String getName();
+    String getDesc();
+}
