+++ src/main/java/com/review/storereview/common/enumerate/UseYN.java
@@ -1,0 +1,5 @@
+package com.review.storereview.common.enumerate;
+
+public enum UseYN {
+    Y, N
+}
