+++ src/main/java/com/review/storereview/controller/BaseController.java
@@ -1,0 +1,4 @@
+package com.review.storereview.controller;
+
+public interface BaseController {
+}
