+++ src/main/java/com/review/storereview/repository/BaseRepository.java
@@ -1,0 +1,4 @@
+package com.review.storereview.repository;
+
+public interface BaseRepository {
+}
