+++ src/main/java/com/review/storereview/service/BaseService.java
@@ -1,0 +1,4 @@
+package com.review.storereview.service;
+
+public interface BaseService {
+}
