+++ src/main/java/com/review/storereview/service/BaseUserService.java
@@ -1,0 +1,8 @@
+package com.review.storereview.service;
+
+import com.review.storereview.dto.impl.request.UserSaveRequestDto;
+
+public interface BaseUserService extends BaseService {
+    // 회원가입
+    String join(UserSaveRequestDto requestDto);
+}
