--- src/test/java/com/review/StoreReview/ReviewServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.review.StoreReview;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ReviewServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
