+++ src/test/java/com/review/storereview/ReviewServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.review.storereview;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ReviewServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
